Apply changes to core
