+++ packages/core/nuxt-theme-module/theme/pages/MyAccount/OrderHistory.vue
@@ -93,7 +93,6 @@ import {
   SfLink
 } from '@storefront-ui/vue';
 import { computed, ref } from '@vue/composition-api';
-
 import { useUserOrders, orderGetters } from '<%= options.generate.replace.composables %>';
 import { AgnosticOrderStatus } from '@vue-storefront/core';
 import { onSSR } from '@vue-storefront/core';
