+++ packages/core/nuxt-theme-module/theme/pages/MyAccount/ShippingDetails.vue
@@ -6,18 +6,20 @@
       :open-tab="1"
       class="tab-orphan"
     >
-      <SfTab data-cy="shipping-details-tab_change" title="Change the address">
+      <SfTab
+        data-cy="shipping-details-tab_change"
+        :title="isNewAddress ? 'Add the address' : 'Update the address'">
         <p class="message">
           Keep your addresses and contact details updated.
         </p>
 
         <ValidationObserver v-slot="{ handleSubmit }">
-          <form id="shipping-details-form" class="form" @submit.prevent="handleSubmit(updateAddress)">
+          <form id="shipping-details-form" class="form" @submit.prevent="handleSubmit(processAddress)">
             <div class="form__horizontal">
               <ValidationProvider rules="required|min:2" v-slot="{ errors }" class="form__element">
                 <SfInput
                   data-cy="shipping-details-input_firstName"
-                  v-model="firstName"
+                  v-model="address.firstName"
                   name="firstName"
                   label="First Name"
                   required
@@ -28,7 +30,7 @@
               <ValidationProvider rules="required|min:2" v-slot="{ errors }" class="form__element">
                 <SfInput
                   data-cy="shipping-details-input_lastName"
-                  v-model="lastName"
+                  v-model="address.lastName"
                   name="lastName"
                   label="Last Name"
                   required
@@ -40,7 +42,7 @@
             <ValidationProvider rules="required|min:5" v-slot="{ errors }" class="form__element">
               <SfInput
                 data-cy="shipping-details-input_streetName"
-                v-model="streetName"
+                v-model="address.streetName"
                 name="streetName"
                 label="Street Name"
                 required
@@ -50,7 +52,7 @@
             </ValidationProvider>
             <SfInput
               data-cy="shipping-details-input_apartment"
-              v-model="apartment"
+              v-model="address.apartment"
               name="apartment"
               label="House/Apartment number"
               required
@@ -60,7 +62,7 @@
               <ValidationProvider rules="required|min:2" v-slot="{ errors }" class="form__element">
                 <SfInput
                   data-cy="shipping-details-input_city"
-                  v-model="city"
+                  v-model="address.city"
                   name="city"
                   label="City"
                   required
@@ -69,9 +71,9 @@
                 />
               </ValidationProvider>
               <ValidationProvider rules="required|min:2" v-slot="{ errors }" class="form__element">
-              <SfInput
-                data-cy="shipping-details-input_state"
-                  v-model="state"
+                <SfInput
+                  data-cy="shipping-details-input_state"
+                  v-model="address.state"
                   name="state"
                   label="State/Province"
                   required
@@ -84,7 +86,7 @@
               <ValidationProvider rules="required|min:4" v-slot="{ errors }" class="form__element">
                 <SfInput
                   data-cy="shipping-details-input_zipCode"
-                  v-model="zipCode"
+                  v-model="address.zipCode"
                   name="zipCode"
                   label="Zip-code"
                   required
@@ -92,22 +94,23 @@
                   :errorMessage="errors[0]"
                 />
               </ValidationProvider>
-              <ValidationProvider :rules="`required|oneOf:${countries.join(',')}`" v-slot="{ errors }" class="form__element">
+              <ValidationProvider
+                :rules="`required|oneOf:${ countries.map(c => c.name).join(',') }`"
+                v-slot="{ errors }"
+                class="form__element">
                 <SfSelect
                   data-cy="shipping-details-select_country"
-                  v-model="country"
+                  v-model="address.country"
                   name="country"
                   label="Country"
                   required
                   :valid="!errors[0]"
-                  :errorMessage="errors[0]"
-                >
+                  :errorMessage="errors[0]">
                   <SfSelectOption
-                    v-for="countryOption in countries"
-                    :key="countryOption"
-                    :value="countryOption"
-                  >
-                    {{ countryOption }}
+                    v-for="{ name, label } in countries"
+                    :key="name"
+                    :value="name">
+                    {{ label }}
                   </SfSelectOption>
                 </SfSelect>
               </ValidationProvider>
@@ -115,7 +118,7 @@
             <ValidationProvider rules="required|min:8" v-slot="{ errors }" class="form__element">
               <SfInput
                 data-cy="shipping-details-input_phoneNumber"
-                v-model="phoneNumber"
+                v-model="address.phoneNumber"
                 name="phone"
                 label="Phone number"
                 required
@@ -123,7 +126,19 @@
                 :errorMessage="errors[0]"
               />
             </ValidationProvider>
-            <SfButton data-cy="shipping-details-btn_update" class="form__button">Update the address</SfButton>
+            <SfCheckbox
+              data-cy="shipping-details-checkbox_isDefault"
+              v-model="address.isDefault"
+              name="isDefault"
+              label="Set as default"
+              class="form__checkbox-isDefault"
+            />
+            <SfButton
+              data-cy="shipping-details-btn_update"
+              class="form__button"
+            >
+              {{ isNewAddress ? 'Add the address' : 'Update the address' }}
+            </SfButton>
           </form>
         </ValidationObserver>
       </SfTab>
@@ -137,23 +152,26 @@
         </p>
         <transition-group tag="div" name="fade" class="shipping-list">
           <div
-            v-for="(shipping, key) in account.shipping"
-            :key="shipping.streetName + shipping.apartment"
-            class="shipping"
-          >
+            v-for="(address, key) in userShippingGetters.getAddresses(shipping)"
+            :key="userShippingGetters.getId(address)"
+            class="shipping">
             <div class="shipping__content">
-              <p class="shipping__address">
-                <span class="shipping__client-name"
-                  >{{ shipping.firstName }} {{ shipping.lastName }}</span
-                ><br />
-                {{ shipping.streetName }} {{ shipping.apartment }}<br />{{
-                  shipping.zipCode
-                }}
-                {{ shipping.city }},<br />{{ shipping.country }}
-              </p>
-              <p class="shipping__address">
-                {{ shipping.phoneNumber }}
-              </p>
+              <div class="shipping__address">
+                <p class="shipping__client-name">
+                  {{ userShippingGetters.getFirstName(address) }}
+                  {{ userShippingGetters.getLastName(address) }}
+                </p>
+                <p>
+                  {{ userShippingGetters.getStreetName(address) }}
+                  {{ userShippingGetters.getApartmentNumber(address) }}
+                </p>
+                <p>
+                  {{ userShippingGetters.getPostCode(address) }}
+                  {{ userShippingGetters.getCity(address) }}
+                </p>
+                <p>{{ countries.find(c => c.name === userShippingGetters.getCountry(address)).label }}</p>
+                <p>{{ userShippingGetters.getPhone(address) }}</p>
+              </div>
             </div>
             <div class="shipping__actions">
               <SfIcon
@@ -163,20 +181,29 @@
                 size="14px"
                 role="button"
                 class="mobile-only"
-                @click="deleteAddress(key)"
+                @click="removeAddress(key)"
               />
-              <SfButton data-cy="shipping-details-btn_change" @click="changeAddress(key)">Change</SfButton>
-              <SfButton data-cy="shipping-details-btn_delete"
+              <SfButton
+                data-cy="shipping-details-btn_change"
+                @click="changeAddress(key)">
+                Change
+              </SfButton>
+
+              <SfButton
+                data-cy="shipping-details-btn_delete"
                 class="shipping__button-delete desktop-only"
-                @click="deleteAddress(key)"
-                >Delete</SfButton
-              >
+                @click="removeAddress(key)">
+                Delete
+              </SfButton>
             </div>
           </div>
         </transition-group>
-        <SfButton data-cy="shipping-details-btn_add" class="action-button" @click="changeAddress(-1)"
-          >Add new address</SfButton
-        >
+        <SfButton
+          data-cy="shipping-details-btn_add"
+          class="action-button"
+          @click="changeAddress(-1)">
+          Add new address
+        </SfButton>
       </SfTab>
     </SfTabs>
   </transition>
@@ -187,10 +214,15 @@ import {
   SfInput,
   SfButton,
   SfSelect,
-  SfIcon
+  SfIcon,
+  SfCheckbox
 } from '@storefront-ui/vue';
+import { getSettings } from '@vue-storefront/commercetools-api';
 import { ValidationProvider, ValidationObserver, extend } from 'vee-validate';
 import { required, min, oneOf } from 'vee-validate/dist/rules';
+import { useUserShipping, userShippingGetters } from '@vsf-enterprise/ct-shipping';
+import { ref, computed } from '@vue/composition-api';
+import { onSSR } from '@vue-storefront/core';
 
 extend('required', {
   ...required,
@@ -215,19 +247,13 @@ export default {
     SfButton,
     SfSelect,
     SfIcon,
+    SfCheckbox,
     ValidationProvider,
     ValidationObserver
   },
-  props: {
-    account: {
-      type: Object,
-      default: () => ({})
-    }
-  },
-  data() {
-    return {
-      editAddress: false,
-      editedAddress: -1,
+  setup() {
+    const resetForm = () => ({
+      id: '',
       firstName: '',
       lastName: '',
       streetName: '',
@@ -237,119 +263,95 @@ export default {
       zipCode: '',
       country: '',
       phoneNumber: '',
-      countries: [
-        'Austria',
-        'Azerbaijan',
-        'Belarus',
-        'Belgium',
-        'Bosnia and Herzegovina',
-        'Bulgaria',
-        'Croatia',
-        'Cyprus',
-        'Czech Republic',
-        'Denmark',
-        'Estonia',
-        'Finland',
-        'France',
-        'Georgia',
-        'Germany',
-        'Greece',
-        'Hungary',
-        'Iceland',
-        'Ireland',
-        'Italy',
-        'Kosovo',
-        'Latvia',
-        'Liechtenstein',
-        'Lithuania',
-        'Luxembourg',
-        'Macedonia',
-        'Malta',
-        'Moldova',
-        'Monaco',
-        'Montenegro',
-        'The Netherlands',
-        'Norway',
-        'Poland',
-        'Portugal',
-        'Romania',
-        'Russia',
-        'San Marino',
-        'Serbia',
-        'Slovakia',
-        'Slovenia',
-        'Spain',
-        'Sweden',
-        'Switzerland',
-        'Turkey',
-        'Ukraine',
-        'United Kingdom',
-        'Vatican City'
-      ]
-    };
-  },
-  methods: {
-    changeAddress(index) {
-      const account = this.account;
-      const shipping = account.shipping[index];
-      if (index > -1) {
-        this.firstName = account.firstName;
-        this.lastName = account.lastName;
-        this.streetName = shipping.streetName;
-        this.apartment = shipping.apartment;
-        this.city = shipping.city;
-        this.state = shipping.state;
-        this.zipCode = shipping.zipCode;
-        this.country = shipping.country;
-        this.phoneNumber = shipping.phoneNumber;
-        this.editedAddress = index;
-      }
-      this.editAddress = true;
-    },
-    updateAddress() {
-      const account = { ...this.account };
-      const shipping = {
-        firstName: this.firstName,
-        lastName: this.lastName,
-        apartment: this.apartment,
-        streetName: this.streetName,
-        city: this.city,
-        state: this.state,
-        zipCode: this.zipCode,
-        country: this.country,
-        phoneNumber: this.phoneNumber
+      isDefault: false
+    });
+
+    const { shipping, load: loadShipping, addAddress, deleteAddress, updateAddress } = useUserShipping();
+    const address = ref(resetForm());
+    const editAddress = ref(false);
+    const editedAddress = ref(-1);
+    const isNewAddress = computed(() => editedAddress.value === -1);
+
+    const getFormById = (index) => {
+      const shippingAddress = userShippingGetters.getAddresses(shipping.value)[index];
+
+      return {
+        id: userShippingGetters.getId(shippingAddress),
+        firstName: userShippingGetters.getFirstName(shippingAddress),
+        lastName: userShippingGetters.getLastName(shippingAddress),
+        streetName: userShippingGetters.getStreetName(shippingAddress),
+        apartment: userShippingGetters.getApartmentNumber(shippingAddress),
+        city: userShippingGetters.getCity(shippingAddress),
+        state: userShippingGetters.getProvince(shippingAddress),
+        zipCode: userShippingGetters.getPostCode(shippingAddress),
+        country: userShippingGetters.getCountry(shippingAddress),
+        phoneNumber: userShippingGetters.getPhone(shippingAddress),
+        isDefault: userShippingGetters.isDefault(shippingAddress)
       };
-      const index = this.editedAddress;
-      if (index > -1) {
-        account.shipping[index] = shipping;
-        this.editedAddress = -1;
-      } else {
-        account.shipping.push(shipping);
-      }
-      this.editAddress = false;
-      this.$emit('update:shipping', account);
-    },
-    deleteAddress(index) {
-      const account = { ...this.account };
-      account.shipping.splice(index, 1);
-      this.$emit('update:shipping', account);
-    }
+    };
+
+    const changeAddress = async (index) => {
+      editedAddress.value = index;
+      address.value = isNewAddress.value ? resetForm() : getFormById(index);
+      editAddress.value = true;
+    };
+
+    const removeAddress = index => deleteAddress(userShippingGetters.getAddresses(shipping.value)[index]);
+
+    const processAddress = async () => {
+      const actionMethod = isNewAddress.value ? addAddress : updateAddress;
+      const { id, ...rest } = address.value;
+
+      await actionMethod({
+        ...(isNewAddress.value ? {} : { id }),
+        ...rest
+      });
+      editAddress.value = false;
+      editedAddress.value = -1;
+    };
+
+    onSSR(async () => {
+      await loadShipping();
+    });
+
+    return {
+      changeAddress,
+      updateAddress,
+      removeAddress,
+      processAddress,
+      userShippingGetters,
+      shipping,
+      editAddress,
+      editedAddress,
+      isNewAddress,
+      countries: getSettings().countries,
+      address
+    };
   }
 };
 </script>
 <style lang='scss' scoped>
 @import '~@storefront-ui/vue/styles';
+
 @mixin for-mobile {
   @media screen and (max-width: $desktop-min) {
     @content;
   }
 }
+
 @mixin for-desktop {
   @media screen and (min-width: $desktop-min) {
     @content;
   }
 }
+
 .form {
+  &__checkbox {
+    &-isDefault {
+      margin-bottom: var(--spacer-2xl);
+    }
+  }
+
   &__element {
     display: block;
     margin-bottom: var(--spacer-2xl);
@@ -378,6 +380,7 @@ export default {
     }
   }
 }
+
 .message {
   margin: 0 0 var(--spacer-2xl) 0;
   font-family: var(--font-family-primary);
@@ -387,16 +390,20 @@ export default {
     font-size: var(--font-base-desktop);
   }
 }
+
 .shipping-list {
   margin-bottom: var(--spacer-2xl);
 }
+
 .shipping {
   display: flex;
   padding: var(--spacer-xl) 0;
   border-top: 1px solid var(--c-light);
+
   &:last-child {
     border-bottom: 1px solid var(--c-light);
   }
+
   &__content {
     flex: 1;
     color: var(--c-text);
@@ -407,6 +414,7 @@ export default {
       font-size: var(--font-sm-desktop);
     }
   }
+
   &__actions {
     flex: 1;
     display: flex;
@@ -419,6 +427,7 @@ export default {
       justify-content: flex-end;
     }
   }
+
   &__button-delete {
     background-color: var(--c-light);
     color: var(--c-text-muted);
@@ -426,29 +435,35 @@ export default {
       margin-left: var(--spacer-xl);
     }
   }
+
   &__address {
-    margin: 0 0 var(--spacer-xl) 0;
-    &:last-child {
+    margin: 0;
+
+    p {
       margin: 0;
     }
   }
+
   &__client-name {
     font-size: var(--font-base-desktop);
     font-weight: 500;
   }
 }
+
 .action-button {
   width: 100%;
   @include for-desktop {
     width: auto;
   }
 }
+
 .tab-orphan {
   @include for-mobile {
     ::v-deep .sf-tabs {
       &__title {
         display: none;
       }
+
       &__content {
         border: 0;
         padding: 0;
